Integrated 'add wrapping conditionals' into the new mutation framework.
